cart work

+++ app/forms/cart_item_form.py
@@ -1,5 +1,5 @@
 from flask_wtf import FlaskForm
-from wtforms import StringField,IntegerField
+from wtforms import StringField,IntegerField, DecimalField
 from wtforms.validators import DataRequired
 from app.models import CartItem
 
@@ -7,4 +7,4 @@ from app.models import CartItem
 class CartItemForm(FlaskForm):
     shoe_id = StringField('shoe_id', validators=[DataRequired()])
     quantity = IntegerField('quantity', validators=[DataRequired()])
-    shoe_size = IntegerField('shoe_size', validators=[DataRequired()])
+    shoe_size =  DecimalField('shoe_size', validators=[DataRequired()])
